api/python_vin_co2/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
motor
httpx
pandas
openpyxl
python-multipart
python-dateutil
